requirements.txt
matplotlib
scikit-learn
scikit-image
opencv-python
yacs
joblib
natsort
h5py
tqdm
einops
gdown
addict
future
lmdb
numpy
pyyaml
requests
scipy
yapf
lpips
packaging
timm==0.4.12
pytest
chardet
yacs
termcolor
submitit
tensorboardX
triton
#causal_conv1d==1.0.0  # causal_conv1d-1.0.0+cu118torch1.13cxx11abiFALSE-cp310-cp310-linux_x86_64.whl
#mamba_ssm==1.0.1  # mamba_ssm-1.0.1+cu118torch1.13cxx11abiFALSE-cp310-cp310-linux_x86_64.whl